Solutions for day 14
Last week was quite busy, so I skipped most of the challenges.
This one was pretty tricky because I was checking the leftover elements
in the wrong order for a while and was just slightly off the mark.

Part 2 is basically a brute force solution (it takes about 6 minutes to
run even with C++), but I'm not going to optimize it now.